testacpi: Speed up the ACPI _PSS test
Before spending a full second measuring the frequency of a CPU at a
given pstate, measure it for 100ms, and pass if we get the expected
result; only fall back to 1s if that fails.  This significantly speeds
up the pstate test, which we run as part of the default test suite.