python: Fix line history navigation with Ctrl-O
Ctrl-O on a line in history submits the line and causes the next call to
readline to start with the following line from history.  That initial
state caused readline to not have a recorded state for the line "after"
all existing history (what would be a fresh new blank line if Ctrl-O had not
left the "vertical cursor" back in history).  Subsequently navigating
down to the end of history would then raise an exception.

Fix that downward navigation to materialize a fresh new blank line when
navigating down from the last line in history.